feat: rexexp to validate text in filter added
