+++ src/main/java/ru/yandex/workshop/main/dto/product/ProductFilter.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import javax.validation.constraints.Pattern;
+import javax.validation.constraints.Size;
 import java.util.List;
 
 @Data
@@ -12,6 +14,8 @@ import java.util.List;
 @AllArgsConstructor
 @Builder
 public class ProductFilter {
+    @Pattern(regexp = "[а-яА-Яa-zA-Z0-9\\s-$]")
+    @Size(min = 3, max = 40, message = "Количество символов от 3 до 40 включительно")
     private String text;
     private List<Long> categories;
     private List<Long> sellerIds;
