--- src/test/java/ru/yandex/workshop/SsaiAppTest.java
@@ -1,13 +1,0 @@
-package ru.yandex.workshop;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-@AutoConfigureTestDatabase
-class SsaiAppTest {
-    @Test
-    void contextLoads() {
-    }
-}
