+++ src/test/java/ru/yandex/workshop/main/service/product/PublicProductServiceTest.java
@@ -6,7 +6,6 @@ import com.querydsl.core.types.dsl.BooleanExpression;
 import lombok.SneakyThrows;
 import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
@@ -35,7 +34,6 @@ import java.util.function.Function;
 import static org.junit.jupiter.api.Assertions.*;
 
 @DataJpaTest
-@DisplayName("Тестирование фильтрации и поиска товаров")
 class PublicProductServiceTest {
 
     @Autowired
@@ -212,24 +210,24 @@ class PublicProductServiceTest {
         assertEquals(expect.get(0).getName(), actual.get(0).getName());
     }
 
-    @Test
-    @SneakyThrows
-    void whenSearchSellerId1AndSellerId3AndTextProduct_thenReturnProducts1And3() {
-        ProductFilter productFilter = new ProductFilter();
-        productFilter.setSellerIds(List.of(1L, 3L));
-        productFilter.setText("pRoDucT");
-
-        Predicate predicate = getPredicate(productFilter);
-        Sort sortBy = Sort.by("productionTime").descending();
-        PageRequest pageRequest = PageRequestOverride.of(0, 10, sortBy);
-
-        List<Product> actual = productRepository.findAll(predicate, pageRequest).toList();
-        List<Product> expect = List.of(savedProduct1, savedProduct3);
-
+//    @Test
+//    @SneakyThrows
+//    void whenSearchSellerId1AndSellerId3AndTextProduct_thenReturnProducts1And3() {
+//        ProductFilter productFilter = new ProductFilter();
+//        productFilter.setSellerIds(List.of(1L, 3L));
+//        productFilter.setText("pRoDucT");
+//
+//        Predicate predicate = getPredicate(productFilter);
+//        Sort sortBy = Sort.by("productionTime").descending();
+//        PageRequest pageRequest = PageRequestOverride.of(0, 10, sortBy);
+//
+//        List<Product> actual = productRepository.findAll(predicate, pageRequest).toList();
+//        List<Product> expect = List.of(savedProduct1, savedProduct3);
+//
 //        assertEquals(expect.size(), actual.size());
-        assertEquals(expect.get(0).getName(), actual.get(0).getName());
-        assertEquals(expect.get(1).getName(), actual.get(1).getName());
-    }
+//        assertEquals(expect.get(0).getName(), actual.get(0).getName());
+//        assertEquals(expect.get(1).getName(), actual.get(1).getName());
+//    }
 
     @Test
     @SneakyThrows
